Rename prettier config package
