+++ apps/admin-ui/package.json
@@ -34,7 +34,6 @@
     "apollo-upload-client": "18.0.1",
     "common": "workspace:*",
     "date-fns": "^4.1.0",
-    "eslint-config-custom": "workspace:*",
     "form-data": "^4.0.3",
     "graphql": "^16.11.0",
     "hds-core": "^4.6.0",
@@ -44,6 +43,7 @@
     "lodash-es": "^4.17.21",
     "next": "15.4.5",
     "next-i18next": "^15.4.2",
+    "prettier-config": "workspace:*",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-error-boundary": "^6.0.0",
