+++ apps/ui/package.json
@@ -22,7 +22,6 @@
     "@sentry/nextjs": "^8.29.0",
     "common": "workspace:*",
     "date-fns": "^4.1.0",
-    "eslint-config-custom": "workspace:*",
     "focus-trap-react": "^11.0.4",
     "graphql": "^16.11.0",
     "hds-core": "^4.6.0",
@@ -33,6 +32,7 @@
     "next-compose-plugins": "^2.2.1",
     "next-i18next": "^15.4.2",
     "nuka-carousel": "^7.0.0",
+    "prettier-config": "workspace:*",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-hook-form": "7.54.2",
