+++ packages/common/package.json
@@ -20,13 +20,12 @@
   },
   "dependencies": {
     "date-fns": "^4.1.0",
-    "eslint-config-custom": "workspace:*",
     "focus-trap-react": "^11.0.4",
     "graphql": "^16.11.0",
     "hds-react": "^4.6.0",
     "i18next": "^25.2.1",
     "lodash-es": "^4.17.21",
-    "next-i18next": "^15.4.2",
+    "prettier-config": "workspace:*",
     "react": "^18.3.1",
     "react-big-calendar": "^1.19.4",
     "react-dom": "^18.3.1",
