+++ packages/prettier-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eslint-config-custom",
+  "name": "prettier-config",
   "license": "MIT",
   "version": "0.0.0",
   "private": true,
